Refactoring

+++ MetaExchange.Models/Order.cs
@@ -13,8 +13,8 @@ namespace MetaExchange.Models
 
         public KindEnum Kind { get; set; }
 
-        public double Amount { get; set; }
+        public decimal Amount { get; set; }
 
-        public double Price { get; set; }
+        public decimal Price { get; set; }
     }
 }
