+++ MetaExchangeServices/Models/RequestInfo.cs
@@ -5,8 +5,8 @@ namespace MetaExchange.Services.Models
     public class RequestInfo
     {
         public OrderTypeEnum OrderType { get; set; }
-        public double BTCAmount { get; set; }
-        public double BTCBalance { get; set; }
-        public double EuroBalance { get; set; }
+        public decimal BTCAmount { get; set; }
+        public decimal BTCBalance { get; set; }
+        public decimal EuroBalance { get; set; }
     }
 }
